Create script to post-process raw dataset
