+++ assets/sunset_info.json
@@ -18,5 +18,9 @@
     {
         "title":"Candilazo",
         "content":"Así se le llama al fenómeno meteorológico en el que las nubes en el cielo muestran una extensa paleta de colores que va del rosa al naranja más intenso."
+    },
+    {
+      "title":"Modo creativo",
+      "content":"Por defecto los colores mostrados en las paletas generadas son representativos considerando todos los colores representativos de todas las imágenes. En el modo creativo las paletas generadas son de los colores representativos de una única imagen."
     }
 ]
